Int casting when returning real_idx

+++ src/kv.c
@@ -46,7 +46,7 @@ int kv_put(kv_t *db, const char *key, const char *value) {
 			char *newval = strdup(value);
 			if (!newval) return -1;
 			entry->value = newval;
-			return real_idx;
+			return (int)real_idx;
 		}
 
 		if (!entry->key || entry->key == (void*)TOMBSTONE) {
@@ -60,7 +60,7 @@ int kv_put(kv_t *db, const char *key, const char *value) {
 			entry->value = newval;
 			entry->key = newkey;
 			db->count++;
-			return real_idx;
+			return (int)real_idx;
 		}	
 	}
 	// if we get to the end, the database is occupied and
